+++ src/app.module.ts
@@ -1,22 +1,14 @@
 import { Module } from '@nestjs/common';
-import { APP_GUARD } from '@nestjs/core';
 
-import { JwtAccessAuthGuard } from './@guards/jwt-access-auth.guard';
 import { AppConfigModule } from './app-config/app-config.module';
 import { AppController } from './app.controller';
 import { AppService } from './app.service';
-import { AuthModule } from './services/auth/auth.module';
-import { UserModule } from './services/user/user.module';
+import { AuthModule } from './modules/auth/auth.module';
+import { UserModule } from './modules/user/user.module';
 
 @Module({
   imports: [UserModule, AppConfigModule, AuthModule],
   controllers: [AppController],
-  providers: [
-    AppService,
-    {
-      provide: APP_GUARD,
-      useClass: JwtAccessAuthGuard,
-    },
-  ],
+  providers: [AppService],
 })
 export class AppModule {}
