+++ src/modules/auth/auth.module.ts
@@ -1,0 +1,30 @@
+import { Module } from '@nestjs/common';
+
+import { UserModule } from '../user/user.module';
+
+import { AuthController } from './auth.controller';
+import { AuthService } from './auth.service';
+import { JwtTokenModule } from '../jwt-token/jwt-token.module';
+import { PassportModule } from '@nestjs/passport';
+import { APP_GUARD } from '@nestjs/core';
+import { JwtAccessAuthGuard } from 'src/@guards/jwt-access-auth.guard';
+import { JwtStrategy } from './strategies/jwt.strategy';
+
+@Module({
+  imports: [
+    UserModule,
+    JwtTokenModule,
+    PassportModule.register({ defaultStrategy: 'jwt' }),
+  ],
+  providers: [
+    AuthService,
+    JwtStrategy,
+    {
+      provide: APP_GUARD,
+      useClass: JwtAccessAuthGuard,
+    },
+  ],
+  controllers: [AuthController],
+  exports: [AuthService],
+})
+export class AuthModule {}
