+++ src/modules/auth/dto/register.dto.ts
@@ -7,7 +7,7 @@ import {
   IsString,
 } from 'class-validator';
 
-import { User } from 'src/services/user/entity/user.entity';
+import { User } from 'src/modules/user/entity/user.entity';
 
 export class UserRegisterDto {
   @ApiProperty({
